i18n on songselector has error
