+++ apps/src/dance/DanceVisualizationColumn.module.scss
@@ -1,0 +1,3 @@
+.selectStyle {
+  width: 100%;
+}
