+++ apps/src/dance/SongSelector.module.scss
@@ -1,0 +1,3 @@
+.selectStyle {
+  width: 100%;
+}
